- Prepare release 1.5.7
- Merge branch 'develop' into release_1_5_7

+++ trade_remedies_api/core/models.py
@@ -15,6 +15,7 @@ from django.contrib.auth.base_user import BaseUserManager, AbstractBaseUser
 from django.contrib.auth.password_validation import validate_password
 from django.contrib.postgres import fields
 from rest_framework.authtoken.models import Token
+from audit.models import Audit
 from security.constants import (
     SECURITY_GROUP_TRA_HEAD_OF_INVESTIGATION,
     SECURITY_GROUPS_TRA,
@@ -366,16 +367,16 @@ class User(AbstractBaseUser, PermissionsMixin, CaseSecurityMixin):
         I.e., no non-draft submissions created.
         Returnes True if the deletion was successful.
         """
-        from audit.models import Audit
 
         user_stats = self.statistics()
         if user_stats.get("non_draft_subs") == 0:
             Audit.objects.filter(created_by_id=self.id).delete()
             self.invitation_set.all().delete()
             organisation = self.organisation.organisation
-            self.organisation.delete()
-            if len(organisation.users) == 1:
-                organisation.delete()
+            if organisation:
+                self.organisation.delete()
+                if len(organisation.users) == 1:
+                    organisation.delete()
             return True
         return False
 
@@ -516,6 +517,11 @@ class User(AbstractBaseUser, PermissionsMixin, CaseSecurityMixin):
         found_role = self.groups.filter(name=role)
         if len(found_role):
             self.groups.remove(found_role[0])
+            if role == SECURITY_GROUP_ORGANISATION_OWNER and not self.groups.all():
+                # A user without a group cannot log on to the public site
+                # when this  function is  used to remove 'organisation owner'
+                # it is safe to grant 'organisation user' instead
+                self.groups.add(Group.objects.get(name=SECURITY_GROUP_ORGANISATION_USER))
         else:
             self.groups.add(Group.objects.get(name=role))
